Added save functionality and a "No Items" view
1. So far, whenever we have been using hard coded values to create the items array and display content on the list. As soon as we kill the app we go back to the default list and none of the changes we made are saved.

To fix this I am using user defaults (as the itemModel is a very simple schema) to save changes made by the user.

3. Lastly I added a view to show when there are no items on the list so the landing page is not black. And also provided a Navigation link to the add Item view where they can add new items.

+++ TodoList/Views/ListView.swift
@@ -12,17 +12,24 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items, content: { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
-                    }
-            })
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List {
+                    ForEach(listViewModel.items, content: { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
+                    })
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+            }
         }
         .navigationTitle("Todo List 📝")
         .toolbar(content: {
